+++ src/components/select/select-textbox/select-textbox.component.js
@@ -1,11 +1,51 @@
-import React from "react";
+import React, { useRef, useLayoutEffect, useMemo } from "react";
 import PropTypes from "prop-types";
+import { createPopper } from "@popperjs/core";
+
 import Textbox from "../../textbox";
+import SelectText from "../__internal__/select-text/select-text.component";
+import guid from "../../../__internal__/utils/helpers/guid/guid";
 import useLocale from "../../../hooks/__internal__/useLocale";
+import useResizeObserver from "../../../hooks/__internal__/useResizeObserver";
+
+const modifiers = [
+  {
+    name: "flip",
+    enabled: false,
+  },
+  {
+    name: "offset",
+    options: {
+      offset: ({ placement, reference }) => {
+        if (placement === "bottom") {
+          return [0, -reference.height];
+        }
+        return [];
+      },
+    },
+  },
+  {
+    name: "sameDimensions",
+    enabled: true,
+    phase: "beforeWrite",
+    requires: ["computeStyles"],
+    fn: ({ state }) => {
+      state.styles.popper.width = `${state.rects.reference.width}px`;
+      state.styles.reference.height = `${state.rects.popper.height}px`;
+    },
+    effect: ({ state }) => {
+      state.elements.popper.style.width = `${state.elements.reference.offsetWidth}px`;
+      state.elements.reference.height = `${state.elements.popper.height}px`;
+    },
+  },
+];
 
 const SelectTextbox = ({
+  accessibilityLabelId = "",
+  "aria-controls": ariaControls,
   value,
   disabled,
+  isOpen,
   readOnly,
   placeholder,
   labelId,
@@ -13,11 +53,54 @@ const SelectTextbox = ({
   onClick,
   onFocus,
   onBlur,
+  onChange,
   selectedValue,
   required,
+  textboxRef,
+  hasTextCursor,
+  transparent,
+  activeDescendantId,
   ...restProps
 }) => {
+  const popperInstance = useRef();
+
+  useLayoutEffect(() => {
+    if (textboxRef && isOpen) {
+      popperInstance.current = createPopper(
+        textboxRef.parentElement.parentElement,
+        textboxRef.parentElement,
+        {
+          strategy: "fixed",
+          modifiers,
+        }
+      );
+    }
+
+    return () => {
+      if (popperInstance.current) {
+        popperInstance.current.destroy();
+        popperInstance.current = null;
+      }
+    };
+  }, [textboxRef, isOpen]);
+
+  const resizeObserverRef = useMemo(
+    () => ({
+      current: textboxRef?.parentElement,
+    }),
+    [textboxRef]
+  );
+
+  useResizeObserver(
+    resizeObserverRef,
+    () => {
+      popperInstance?.current?.update();
+    },
+    !isOpen
+  );
+
   const l = useLocale();
+  const textId = useRef(guid());
 
   function handleTextboxClick(event) {
     if (disabled || readOnly) {
@@ -45,7 +128,9 @@ const SelectTextbox = ({
 
   function getTextboxProps() {
     return {
-      placeholder: placeholder || l.select.placeholder(),
+      placeholder: hasTextCursor
+        ? placeholder || l.select.placeholder()
+        : undefined,
       disabled,
       readOnly,
       required,
@@ -53,8 +138,43 @@ const SelectTextbox = ({
       onFocus: handleTextboxFocus,
       onBlur: handleTextboxBlur,
       labelId,
+      type: "text",
       ...restProps,
     };
+  }
+
+  function getInputAriaAttributes() {
+    return {
+      "aria-expanded": isOpen,
+      "aria-labelledby": hasTextCursor
+        ? `${labelId} ${accessibilityLabelId}`
+        : `${labelId} ${textId.current}`,
+      "aria-activedescendant": activeDescendantId,
+      "aria-controls": ariaControls,
+      "aria-autocomplete": hasTextCursor ? "both" : undefined,
+      role: readOnly ? undefined : "combobox",
+    };
+  }
+
+  function renderSelectText() {
+    if (hasTextCursor) {
+      return null;
+    }
+
+    return (
+      <SelectText
+        textId={textId.current}
+        transparent={transparent}
+        onKeyDown={handleSelectTextKeydown}
+        {...getTextboxProps()}
+      />
+    );
+  }
+
+  function handleSelectTextKeydown(event) {
+    if (event.key.length === 1) {
+      onChange({ target: { value: event.key } });
+    }
   }
 
   return (
@@ -63,13 +183,24 @@ const SelectTextbox = ({
       inputIcon="dropdown"
       autoComplete="off"
       size={size}
+      onChange={onChange}
       value={selectedValue}
+      {...getInputAriaAttributes()}
       {...getTextboxProps()}
-    />
+    >
+      {renderSelectText()}
+    </Textbox>
   );
 };
 
 const formInputPropTypes = {
+  /**
+   * Id of the element containing the currently displayed value
+   * to be read by voice readers
+   * @private
+   * @ignore
+   */
+  accessibilityLabelId: PropTypes.string,
   /** Id attribute of the input element */
   id: PropTypes.string,
   /** Name attribute of the input element */
@@ -97,6 +228,12 @@ const formInputPropTypes = {
   labelWidth: PropTypes.number,
   /** Width of an input in percentage. Works only when labelInline is true */
   inputWidth: PropTypes.number,
+  /**
+   * @ignore
+   * @private
+   * If true, the select is open
+   */
+  isOpen: PropTypes.bool,
   /** Size of an input */
   size: PropTypes.oneOf(["small", "medium", "large"]),
   /** Placeholder string to be displayed in input */
@@ -120,10 +257,21 @@ const formInputPropTypes = {
 SelectTextbox.propTypes = {
   ...formInputPropTypes,
   /**
+   * @ignore
+   * @private
+   * Id attribute of the select list
+   */
+  "aria-controls": PropTypes.string,
+  /**
    * @private
    * @ignore
    * Value to be displayed in the Textbox */
   formattedValue: PropTypes.string,
+  /**
+   * @private
+   * @ignore
+   * If true, the input will be displayed */
+  hasTextCursor: PropTypes.bool,
   /**
    * @private
    * @ignore