+++ src/hooks/__internal__/useScrollBlock/index.d.ts
@@ -1,0 +1,4 @@
+export default function useScrollBlock(): {
+  blockScroll: () => void;
+  allowScroll: () => void;
+};